requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Flask==2.2.0
Flask-Cors==3.0.10
gunicorn==20.1.0
h11==0.13.0
html2image==2.0.1
idna==3.3
image-slicer==2.1.1
imgkit==1.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
Pillow==7.2.0
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyppeteer==1.0.2
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
urllib3==1.26.11
webdriver-manager==3.8.3
websockets==10.3
Werkzeug==2.2.1
wsproto==1.1.0
zipp==3.8.1
